fix: move Perplexity memory loading to API route (avoids client bundle issue)
Moved loadPerplexityMemories from writer.ts to chat/route.ts to avoid
Next.js bundler issues with fs/path in client-side modules. Memories
are now injected into systemInstructions at the API route level.
